add three functrions
Function to asses the number of teachers in which one of the levels.

This functions works considering the possibility that are teachers in each one of these levels.

isced_counter is a function to read in which levels they are.

required_sample is how many teachers we will need in each school.